* 'master' of github.com:levelsix/mobsters-client:
  Bugfixes
  Skills: Roid Rage, Momentum, manual activation
  Bugfixes
  Posion improvements
  Skills: poison part 2
  Skills: poison part 1

+++ Utopia/Classes/Battle/Skills/SkillBombs.m
@@ -214,7 +214,7 @@ static const NSInteger bombsMaxSearchIterations = 1000;
     
     // Deal damage
     [battleLayer performAfterDelay:delay block:^{
-      battleLayer.enemyDamageDealt = totalDamage;
+      battleLayer.enemyDamageDealt = (int)totalDamage;
       [battleLayer dealDamage:(int)totalDamage enemyIsAttacker:YES usingAbility:YES withTarget:nil withSelector:nil];
     }];
